Extract out command interface and service modules
